+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed: Firing sequence of `common_toolkit_loaded` hook
 - Added: `ctk_environment` filter
+- Added: Ability to cache JSON config file
 - Added: Ability to disable WordPress core, plugin and/or theme updates
 - Added: Ability to modify or disable WordPress heartbeat
 - Added: Ability to disable WordPress search